Move Section.database.functions to Section.controllers
