+++ src/routes/Section.routes.ts
@@ -2,23 +2,21 @@ import {
   NextFunction, Request, Response, Router,
 } from 'express';
 import Controller from '../controllers/RouterWrapper.interface';
-import { createSections } from '../controllers/Project.controllers';
 import UpdateSectionDto from '../dtos/Sections/UpdateSection.dto';
-import ProjectNotFoundException from '../exceptions/projects/ProjectNotFoundException';
 import SectionNotFoundException from '../exceptions/sections/SectionNotFoundException';
 import validationMiddleware from '../middleware/validation.middleware';
 import {
   createTask,
   deleteSection,
   getSection,
-  getSections,
   updateSection,
-} from '../database_functions/Section.database.functions';
+} from '../controllers/Section.controllers';
 import AssigneeNotInProjectException from '../exceptions/projects/AssigneeNotInProjectException';
 import CreateTaskDto from '../dtos/Sections/CreateTask.dto';
 import { getTasks } from '../database_functions/Task.database.functions';
 import authMiddleware from '../middleware/auth.middleware';
 import RequestWithUser from '../interfaces/RequestWithUser.interface';
+import { sectionPermissionMiddleware } from '../middleware/permission.middleware';
 
 class SectionRoutes implements Controller {
   public path = '/sections';
@@ -33,27 +31,37 @@ class SectionRoutes implements Controller {
     this.router.all(`${this.path}`, authMiddleware);
     this.router.all(`${this.path}/*`, authMiddleware);
 
-    this.router.get(`${this.path}`, SectionRoutes.getSections);
+    this.router.get(
+      `${this.path}/:sectionId`,
+      sectionPermissionMiddleware('get'),
+      SectionRoutes.getSection,
+    );
 
-    this.router.get(`${this.path}/:sectionId`, SectionRoutes.getSection);
-
-    this.router.get(`${this.path}/:sectionId/tasks`, SectionRoutes.getTasks);
-
-    this.router.post(`${this.path}`, SectionRoutes.createSection);
+    this.router.get(
+      `${this.path}/:sectionId/tasks`,
+      sectionPermissionMiddleware('get'),
+      SectionRoutes.getTasks,
+    );
 
     this.router.post(
       `${this.path}/:sectionId/tasks`,
       validationMiddleware(CreateTaskDto),
+      sectionPermissionMiddleware('update'),
       SectionRoutes.createTask,
     );
 
     this.router.patch(
       `${this.path}/:sectionId`,
       validationMiddleware(UpdateSectionDto),
+      sectionPermissionMiddleware('update'),
       SectionRoutes.updateSection,
     );
 
-    this.router.delete(`${this.path}/:sectionId`, SectionRoutes.deleteSection);
+    this.router.delete(
+      `${this.path}/:sectionId`,
+      sectionPermissionMiddleware('delete'),
+      SectionRoutes.deleteSection,
+    );
 
     this.router.all(`${this.path}`, authMiddleware);
     this.router.all(`${this.path}/*`, authMiddleware);
@@ -83,9 +91,7 @@ class SectionRoutes implements Controller {
   private static async getSection(req: Request, res: Response, next: NextFunction) {
     const { sectionId } = req.params;
 
-    const { user } = req as RequestWithUser;
-
-    const sectionsOrError = await getSection(user._id.toString(), sectionId);
+    const sectionsOrError = await getSection(sectionId);
     if (sectionsOrError.type === 'error') {
       next(new SectionNotFoundException(sectionId));
       return;
@@ -94,24 +100,6 @@ class SectionRoutes implements Controller {
     res
       .json({
         section: sectionsOrError.data,
-      })
-      .end();
-  }
-
-  private static async getSections(req: Request, res: Response, next: NextFunction) {
-    const { projectId } = req.body;
-
-    const sectionsOrError = await getSections(projectId);
-    if (sectionsOrError.type === 'error') {
-      next(new ProjectNotFoundException(projectId));
-      return;
-    }
-
-    const sections = sectionsOrError.data;
-
-    res
-      .json({
-        sections,
       })
       .end();
   }
@@ -165,29 +153,6 @@ class SectionRoutes implements Controller {
         task,
       })
       .end();
-  }
-
-  private static async createSection(req: Request, res: Response, next: NextFunction) {
-    const { projectId } = req.params;
-    const { name = '', colour = '', icon = '' } = req.body;
-
-    const sections = await createSections(projectId, [
-      {
-        name,
-        colour,
-        icon,
-      },
-    ]);
-
-    if (!sections) {
-      next(new ProjectNotFoundException(projectId));
-    } else {
-      res
-        .json({
-          section: sections[0],
-        })
-        .end();
-    }
   }
 
   private static async deleteSection(req: Request, res: Response, next: NextFunction) {
